requirements.txt
scapy>=2.5.0
rich>=13.7.0
Flask>=3.0.0
gunicorn>=21.2.0
Werkzeug>=3.0.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
